requirements.txt
streamlit>=1.36
pymongo[srv]>=4.6
google-generativeai>=0.7
python-dotenv>=1.0
sentence-transformers>=3.0
html2text>=2024.2.26
beautifulsoup4>=4.12
lxml>=5.2
regex>=2024.7.24
